[WIP] Client-side recreation

+++ Assets/Scripts/TCPToolkit/TCPToolkit.cs
@@ -26,10 +26,10 @@ namespace ubv
                 {
                     Packet packet = new Packet(new byte[NET_PROTOCOL_ID.Length + data.Length]);
                     int index = 0;
-                    for (ushort i = 0; i < 4; i++, index++)
+                    for (int i = 0; i < 4; i++, index++)
                         packet.RawBytes[index] = NET_PROTOCOL_ID[i];
 
-                    for (ushort i = 0; i < data.Length; i++, index++)
+                    for (int i = 0; i < data.Length; i++, index++)
                         packet.RawBytes[index] = data[i];
 
                     return packet;
